@@ -4,22 +4,25 @@ import PushBullet from 'pushbullet';
 import striptags from 'striptags';
 
 // Send the push notification.
-// Todo: why bother with async / await at all ?
 export default {
   async sendPushBulletNotification(config, feedData) {
     const pusher = new PushBullet(config.accessToken);
-    await pusher.link(
-      {},
-      feedData.title,
-      feedData.link,
-      striptags(feedData.description),
-      (error, response) => {
-        if (error) {
-          return false;
+
+    return new Promise((resolve, reject) => {
+      pusher.link(
+        {},
+        feedData.title,
+        feedData.link,
+        striptags(feedData.description),
+        (error, response) => {
+          if (error) {
+            resolve(false);
+          } else {
+            // console.log('Push notification sent successfully!');
+            resolve(response);
+          }
         }
-        // console.log('Push notification sent successfully!');
-        return response;
-      }
-    );
+      );
+    });
   },
 }
